# openapi.yaml
openapi: 3.0.3
info:
  title: AuthOS API
  description: |
    Comprehensive Laravel 12 authentication service API - An Auth0 alternative providing OAuth 2.0, OpenID Connect, multi-factor authentication, and single sign-on capabilities.
    
    ## Features
    - OAuth 2.0 & OpenID Connect server implementation
    - Multi-tenant organization management
    - Role-based access control (RBAC)
    - Multi-factor authentication (MFA) with TOTP
    - Real-time authentication monitoring
    - Comprehensive user and application management
    - Rate limiting with role-based adjustments
    
    ## Authentication
    This API uses OAuth 2.0 Bearer tokens for authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer {access_token}
    ```
    
    ## Rate Limiting
    Different endpoints have varying rate limits based on their usage patterns:
    - Authentication: 10 requests/minute (per IP)
    - Registration: 5 requests/hour (per IP) 
    - Standard API: 1000 requests/hour (per user)
    - Admin API: 200 requests/hour (per user)
    - Bulk Operations: 100 requests/hour (per user)
    - MFA: 20 requests/hour (per user)
    - OAuth: 20 requests/minute (per IP)
    
    Rate limits are increased for users with higher roles (Super Admin gets 5x, Organization Admin gets 3x, Application Admin gets 2x).
    
    ## Error Handling
    All API errors follow a consistent JSON format with appropriate HTTP status codes.
    
  version: 1.0.0
  contact:
...
servers:
...
paths:
...
  /auth/user:
    get:
      tags:
        - Authentication
      summary: Get current user
      description: Retrieve authenticated user information based on token scopes
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        200:
          description: Current user information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
        401:
          $ref: '#/components/responses/UnauthorizedError'
...
  /auth/revoke:
    post:
      tags:
        - Authentication
      summary: Revoke access token
      description: Revoke the current access token
      operationId: revokeToken
      security:
        - bearerAuth: []
      responses:
        200:
          description: Token revoked successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Token revoked successfully"

  # OAuth 2.0 Endpoints
...
  /profile:
    get:
      tags:
        - Profile Management
      summary: Get current user profile
      description: Get the authenticated user's profile information
      operationId: getProfile
      security:
        - bearerAuth: []
      responses:
        200:
          description: Profile retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/UserProfile'

    put:
      tags:
        - Profile Management
      summary: Update profile
      description: Update the authenticated user's profile
      operationId: updateProfile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  maxLength: 255
                profile:
                  type: object
                  properties:
                    timezone:
                      type: string
                      example: "America/New_York"
                    language:
                      type: string
                      enum: [en, es, fr, de]
                      example: "en"
                    bio:
                      type: string
                      maxLength: 1000
                    location:
                      type: string
                      maxLength: 255
                    website:
                      type: string
                      format: uri
      responses:
        200:
          description: Profile updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/UserProfile'
                  message:
                    type: string
                    example: "Profile updated successfully"
...
components:
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
          format: email
        avatar:
          type: string
          format: uri
          nullable: true
        profile:
          type: object
        organization:
          $ref: '#/components/schemas/Organization'
        mfa_enabled:
          type: boolean
        preferences:
          type: object
          properties:
            notifications:
              type: object
              properties:
                email:
                  type: boolean
                push:
                  type: boolean
            theme:
              type: string
              enum: [light, dark, auto]
    UserInfo:
      type: object
      description: "OpenID Connect UserInfo response (scope-dependent)"
      properties:
        sub:
          type: string
          description: "Subject identifier (user ID)"
          example: "1"
        name:
          type: string
          description: "Full name (profile scope)"
          example: "John Doe"
        email:
          type: string
          format: email
          description: "Email address (email scope)"
          example: "john@example.com"
        email_verified:
          type: boolean
          description: "Email verification status (email scope)"
          example: true
        profile:
          type: object
          description: "Profile information (profile scope)"
        picture:
          type: string
          format: uri
          description: "Avatar URL (profile scope)"
        updated_at:
          type: integer
          description: "Last updated timestamp"
    Organization:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "TechCorp Solutions"
        slug:
          type: string
          example: "techcorp-solutions"
        settings:
          type: object
          example: {"require_mfa": false, "session_timeout": 1440}
        is_active:
          type: boolean
          example: true
        applications_count:
          type: integer
          description: "Number of applications in organization"
        users_count:
          type: integer
          description: "Number of users with access to organization applications"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: "Error code"
        error_description:
          type: string
          description: "Human-readable error description"
        details:
          type: object
          description: "Additional error details"
...
  responses:
    UnauthorizedError:
      description: "Authentication required"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "unauthorized"
            error_description: "Authentication required"
      headers:
        WWW-Authenticate:
          schema:
            type: string
          example: "Bearer"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        OAuth 2.0 Bearer token authentication. Include the access token in the Authorization header:
        ```
        Authorization: Bearer {access_token}
        ```
        
        ### Obtaining Access Tokens
        Access tokens can be obtained through:
        - **User Registration/Login**: `/auth/register` or `/auth/login`
        - **OAuth Authorization Code Flow**: `/oauth/authorize` → `/oauth/token`
        - **Client Credentials Flow**: `/oauth/token` (for server-to-server)
        
        ### Token Scopes
        Tokens may include different scopes that determine access levels:
        - `openid`: Basic OpenID Connect access
        - `profile`: Access to user profile information
        - `email`: Access to user email address
        - `read`: Read access to resources
        - `write`: Write access to resources
        - `admin`: Administrative access (for management endpoints)
        
        ### Token Lifetime
        Access tokens typically expire after 1 hour. Use refresh tokens to obtain new access tokens without re-authentication.
security:
  - bearerAuth: []
...